requirements.txt
flask
langchain
langchain-openai
langchain-milvus
openai
python-dotenv
pymilvus